refactor: move chat attachment helpers to ipc utils
