+++ src/ipc/utils/chat_attachment_utils.ts
